Refactor interactive pty helpers to dedicated file
Extract generic RunCommandInteractive and WaitForPromptAndRespond helpers
into interactive.go, simplifying resume_test.go. Documents the critical
ACCESSIBLE=1 requirement for huh library to read from pty stdin.

Entire-Checkpoint: ceaea776cdb0
