+++ cmd/entire/cli/integration_test/resume_test.go
@@ -3,9 +3,7 @@
 package integration
 
 import (
-	"bytes"
 	"fmt"
-	"io"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -16,7 +14,6 @@ import (
 
 	"entire.io/cli/cmd/entire/cli/strategy"
 
-	"github.com/creack/pty"
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
 )
@@ -511,73 +508,10 @@ func (env *TestEnv) RunResumeForce(branchName string) (string, error) {
 
 // RunResumeInteractive executes the resume command with a pty, allowing
 // interactive prompt responses. The respond function receives the pty for
-// reading output and writing input. Timeouts and cleanup are managed centrally.
-// The respond function should read from the pty to find prompts and write responses.
-// All output read by respond is captured and returned.
+// reading output and writing input. See RunCommandInteractive for details.
 func (env *TestEnv) RunResumeInteractive(branchName string, respond func(ptyFile *os.File) string) (string, error) {
 	env.T.Helper()
-
-	cmd := exec.Command(getTestBinary(), "resume", branchName)
-	cmd.Dir = env.RepoDir
-	cmd.Env = append(os.Environ(),
-		"ENTIRE_TEST_CLAUDE_PROJECT_DIR="+env.ClaudeProjectDir,
-		"TERM=xterm",
-		"ACCESSIBLE=1", // Use accessible mode for simpler text prompts
-	)
-
-	// Start command with a pty
-	ptmx, err := pty.Start(cmd)
-	if err != nil {
-		return "", fmt.Errorf("failed to start pty: %w", err)
-	}
-	defer ptmx.Close()
-
-	// Let the respond function interact with the pty and collect output
-	var respondOutput string
-	respondDone := make(chan struct{})
-	go func() {
-		defer close(respondDone)
-		respondOutput = respond(ptmx)
-	}()
-
-	// Wait for respond function with timeout
-	select {
-	case <-respondDone:
-		// respond completed
-	case <-time.After(10 * time.Second):
-		env.T.Log("Warning: respond function timed out")
-	}
-
-	// Collect any remaining output after respond is done
-	var remaining bytes.Buffer
-	remainingDone := make(chan struct{})
-	go func() {
-		defer close(remainingDone)
-		_, _ = io.Copy(&remaining, ptmx)
-	}()
-
-	// Wait for process to complete with timeout
-	cmdDone := make(chan error, 1)
-	go func() {
-		cmdDone <- cmd.Wait()
-	}()
-
-	var cmdErr error
-	select {
-	case cmdErr = <-cmdDone:
-		// process completed
-	case <-time.After(10 * time.Second):
-		_ = cmd.Process.Kill()
-		cmdErr = fmt.Errorf("process timed out")
-	}
-
-	// Give remaining output goroutine time to finish after process exits
-	select {
-	case <-remainingDone:
-	case <-time.After(1 * time.Second):
-	}
-
-	return respondOutput + remaining.String(), cmdErr
+	return env.RunCommandInteractive([]string{"resume", branchName}, respond)
 }
 
 // GitMerge merges a branch into the current branch.
@@ -756,32 +690,6 @@ func TestResume_LocalLogNewerTimestamp_ForceOverwrites(t *testing.T) {
 	}
 }
 
-// waitForPromptAndRespond reads from the pty until it sees the expected prompt text,
-// then writes the response. Returns the output read so far.
-func waitForPromptAndRespond(ptyFile *os.File, promptSubstring, response string, timeout time.Duration) (string, error) {
-	var output bytes.Buffer
-	buf := make([]byte, 1024)
-	deadline := time.Now().Add(timeout)
-
-	for time.Now().Before(deadline) {
-		// Set read deadline to avoid blocking forever
-		_ = ptyFile.SetReadDeadline(time.Now().Add(100 * time.Millisecond))
-		n, err := ptyFile.Read(buf)
-		if n > 0 {
-			output.Write(buf[:n])
-			if strings.Contains(output.String(), promptSubstring) {
-				// Found the prompt, send response
-				_, _ = ptyFile.WriteString(response)
-				return output.String(), nil
-			}
-		}
-		if err != nil && !os.IsTimeout(err) {
-			return output.String(), err
-		}
-	}
-	return output.String(), fmt.Errorf("timeout waiting for prompt containing %q", promptSubstring)
-}
-
 // TestResume_LocalLogNewerTimestamp_UserConfirmsOverwrite tests that when the user
 // confirms the overwrite prompt interactively, the local log is overwritten.
 func TestResume_LocalLogNewerTimestamp_UserConfirmsOverwrite(t *testing.T) {
@@ -823,8 +731,7 @@ func TestResume_LocalLogNewerTimestamp_UserConfirmsOverwrite(t *testing.T) {
 
 	// Resume interactively and confirm the overwrite
 	output, err := env.RunResumeInteractive(featureBranch, func(ptyFile *os.File) string {
-		// Wait for the accessible prompt "[y/N]", then send 'y'
-		out, promptErr := waitForPromptAndRespond(ptyFile, "[y/N]", "y\n", 10*time.Second)
+		out, promptErr := WaitForPromptAndRespond(ptyFile, "[y/N]", "y\n", 10*time.Second)
 		if promptErr != nil {
 			t.Logf("Warning: %v", promptErr)
 		}
@@ -888,8 +795,7 @@ func TestResume_LocalLogNewerTimestamp_UserDeclinesOverwrite(t *testing.T) {
 
 	// Resume interactively and decline the overwrite
 	output, err := env.RunResumeInteractive(featureBranch, func(ptyFile *os.File) string {
-		// Wait for the accessible prompt "[y/N]", then send 'n'
-		out, promptErr := waitForPromptAndRespond(ptyFile, "[y/N]", "n\n", 10*time.Second)
+		out, promptErr := WaitForPromptAndRespond(ptyFile, "[y/N]", "n\n", 10*time.Second)
 		if promptErr != nil {
 			t.Logf("Warning: %v", promptErr)
 		}
